+++ engine/app/components/citizens_advice_components/icons/arrow_up_down.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module CitizensAdviceComponents
+  module Icons
+    class ArrowUpDown < CitizensAdviceComponents::Icons::Base
+      def icon_name
+        "arrow-up-down"
+      end
+    end
+  end
+end