+++ engine/app/components/citizens_advice_components/icons/plus_minus.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module CitizensAdviceComponents
+  module Icons
+    class PlusMinus < CitizensAdviceComponents::Icons::Base
+      def icon_name
+        "plus-minus"
+      end
+    end
+  end
+end