+++ engine/previews/citizens_advice_components/icons_preview.rb
@@ -3,5 +3,9 @@
 module CitizensAdviceComponents
   class IconsPreview < ViewComponent::Preview
     def icons; end
+
+    def arrow_up_down; end
+
+    def plus_minus; end
   end
 end